Update WireMock to version 3.6.0 to allow use of the start() method
WireMock 3.6.0 provided the start() and stop() extension lifecycle
methods which can be used to initialise extensions when WireMock starts
and allow them to perform and cleanup when WireMock shuts down.

This plugin used a hack to initialise the jwt so this change removes the
hack in favour of the new start() method
