+++ build.gradle
@@ -22,7 +22,7 @@ repositories {
 
 ext {
   versions = [
-      wiremock    : "3.5.2",
+      wiremock    : "3.6.0",
       jwt         : '4.4.0',
       jose4j      : '0.9.6',
   ]
